✨ feat: 리크루팅 API 연결 타입 일치
