+++ src/components/recruting/_02_prepare/_01/RecruitingCalenderPicker.tsx
@@ -2,7 +2,7 @@ import "../../../../style/calendar.css";
 import dayGridPlugin from "@fullcalendar/daygrid";
 import interactionPlugin from "@fullcalendar/interaction";
 import FullCalendar from "@fullcalendar/react";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { DateSelectArg, EventClickArg } from "@fullcalendar/core";
 import {
   CALENDAR_ITEMS,
@@ -41,6 +41,25 @@ export default function RecrutingCalenderPicker() {
   );
   const [editMode, setEditMode] = useState(false);
 
+  const [recruitSchedules, setRecruitSchedules] = useState({
+    stage1Start: "",
+    stage1End: "",
+    stage2Start: "",
+    stage2End: "",
+    stage3Start: "",
+    stage3End: "",
+    stage4Start: "",
+    stage4End: "",
+    stage5Start: "",
+    stage5End: "",
+    stage6Start: "",
+    stage6End: "",
+    stage7Start: "",
+    stage7End: "",
+    stage8Start: "",
+    stage8End: ""
+  });
+
   // 날짜 선택 시 이벤트 추가 함수
   const handleDateSelect = (selectInfo: DateSelectArg) => {
     const title = currrentTitle;
@@ -67,7 +86,17 @@ export default function RecrutingCalenderPicker() {
       };
       setEvents((prevEvents) => [...prevEvents, newEvent]);
       setCompletedTitles((prevTitles) => [...prevTitles, title]); // 완료된 제목에 추가
-      console.log(events);
+
+      // 각 단계에 대한 일정 저장
+      const stageNumber = CALENDAR_ITEMS.indexOf(title) + 1;
+      setRecruitSchedules((prev) => ({
+        ...prev,
+        [`stage${stageNumber}Start`]: selectInfo.startStr,
+        [`stage${stageNumber}End`]: addDays(new Date(selectInfo.endStr), 0)
+          .toISOString()
+          .split("T")[0]
+      }));
+
       setInstructionMessage("");
       // title이 '면접 기간'인 경우에만 전역 상태에 시작, 종료 날짜 저장
       if (title === "면접 기간") {
@@ -145,6 +174,20 @@ export default function RecrutingCalenderPicker() {
   const handleCalendarMouseLeave = () => {
     setInstructionMessage(null);
   };
+
+  const checkAllStagesSelected = () => {
+    const allStagesSelected = CALENDAR_ITEMS.every((item) =>
+      completedTitles.includes(item)
+    );
+    if (allStagesSelected) {
+      console.log("recruitSchedules:", recruitSchedules);
+      // 여기에서 recruitSchedules를 원하는 대로 처리할 수 있습니다 (예: API로 전송)
+    }
+  };
+
+  useEffect(() => {
+    checkAllStagesSelected();
+  }, [completedTitles]);
 
   return (
     <div className="mt-[30px] pl-4 bg-white-100">
